// frontend/src/api/backend/backendSchemas.ts
export type ObjectId = {};
// ...
export type RoleResponseDto = {
  /**
   * Unique identifier of the role
   *
   * @example 65def325de3deddff04f79a0
   */
  _id: ObjectId;
  /**
   * Name of the role
   *
   * @example Super Admin
   */
  name: string;
  /**
   * Code of the role
   *
   * @example super-admin
   */
  code: string;
  /**
   * Description of the role
   *
   * @example Super Admin
   */
  description: string;
  /**
   * Array of module IDs associated with the role
   *
   * @example 65e3579de0ca70fe8f2e148e
   * @example 65e357e0e0ca70fe8f2e1490
   * @example 65def137de3deddff04f799b
   */
  moduleId: ObjectId[];
  /**
   * Permissions associated with the role
   */
  permissions: PermissionDto[];
  /**
   * Status of the role
   *
   * @example true
   */
  status: boolean;
};

export type DepartmentsDto = {
  /**
   * Unique identifier of the department
   *
   * @example 65e767af49834fc6088bfba7
   */
  _id: ObjectId;
  /**
   * Code of the department
   *
   * @example chief-department
   */
  code: string;
  /**
   * Name of the department
   *
   * @example Chief Department
   */
  name: string;
  /**
   * Array of roles associated with the department
   */
  roles: RoleResponseDto[];
};

export type UserResponseDto = {
  /**
   * Unique identifier for the user
   *
   * @example 677d567870dc1af598b322a7
   */
  _id: ObjectId;
  /**
   * Name of the user
   *
   * @example Adamtest2
   */
  name: string;
  /**
   * Email address of the user
   *
   * @example vrameshbapu1@nrl.com.au
   */
  email: string;
  /**
   * List of departments the user is associated with
   */
  departments: DepartmentsDto[];
  /**
   * Status of the user account
   *
   * @example true
   */
  status: boolean;
  /**
   * User who created this record
   *
   * @example bbharathivel@nrl.com.au
   */
  createdBy: string;
  /**
   * User who last updated this record
   *
   * @example bbharathivel@nrl.com.au
   */
  updatedBy: string;
};